major reorganization of the file structure: all headers now in LHPC/include/, all source files in LHPC/source/
major reorganization of the file structure: all headers now in
LHPC/include/, all source files in LHPC/source/. relative paths between
headers unchanged. will break existing code, but only in so far as the
include path for the compiler has to be changed. BOLlib is included but
is now independent.